Boot via boot.scr, too, at the end.
If we don't boot{m,z,?} from boot.scr the setup for u-boot will
continue and try to load kernel and FDT from "/boot/<file>" which
will fail because it is now the boot-partition and our boot files
are accessible as "/<file>" instead.
Also, we need to specify the initrd.